backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = {text = "ITM"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "openai (>=2.26.0,<3.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
]


[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
